+++ packages/capillary-ui/src/Components/lineinputs/LabeledInputControl.ts
@@ -32,24 +32,5 @@ export abstract class LabeledInputControl<
             white-space: nowrap;
         }
 
-        &:has(> input:required:invalid:not(:disabled):not([readonly]):not([aria-invalid="true"]))::after {
-            content: "!";
-            position: absolute;
-            z-index: 3;
-            inset-block-start: -.5em;
-            inset-inline-end: -.5em;
-            display: grid;
-            inline-size: 1em;
-            block-size: 1em;
-            place-items: center;
-            color: var(--required-indicator-color);
-            background: var(--required-indicator-background);
-            border-radius: 50%;
-            font-size: .75em;
-            font-weight: 700;
-            line-height: 1;
-            pointer-events: none;
-        }
-
     `
 }